Merged PR 89: Add email and password authentication
In addition to the existing OAuth login using Google, this update adds an email and password based on authentication by allowing users to create accounts, login with them, and reset password to the account.

This PR also updates Angular (and Angular Material) from version 11 to 13, AngularFire from 5 to 7, Firebase from 8 to 9, and Ionic from 5 to 6.